% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortRelationship.R
\docType{package}
\name{CohortRelationship-package}
\alias{CohortRelationship}
\alias{CohortRelationship-package}
\title{CohortRelationship: An R package to compute relationships between set of instantiated cohorts using interval algebra.}
\description{
An R package with a Shiny viewer to explore profiles of patients in a cohort.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortRelationship/}
  \item \url{https://github.com/OHDSI/CohortRelationship}
  \item Report bugs at \url{https://github.com/OHDSI/CohortRelationship/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
