% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Private.R
\name{createIfNotExist}
\alias{createIfNotExist}
\title{Create a Folder if It Does Not Exist}
\usage{
createIfNotExist(type, name, recursive = TRUE, errorMessage = NULL)
}
\arguments{
\item{type}{A character string specifying the type of entity to create. Currently, only \code{"folder"} is supported.}

\item{name}{A character string specifying the name or path where the folder should be created.}

\item{recursive}{A logical value indicating whether to create the directory recursively (i.e., creating any necessary parent directories). Defaults to \code{TRUE}.}

\item{errorMessage}{An optional AssertCollection object (from the \code{checkmate} package) used to collect error messages. If \code{NULL} or not a valid AssertCollection, a new one is created.}
}
\value{
Invisibly returns the AssertCollection object containing any collected error messages.
}
\description{
This function creates a folder (directory) at a specified location if it does not already exist.
It also verifies that the folder exists and is accessible by asserting its execute permissions.
Any issues or error messages can be collected using an AssertCollection from the \code{checkmate} package.
}
\details{
The function first ensures that an appropriate AssertCollection object is available. If \code{type} is not \code{NULL}, it checks if
the \code{name} parameter is non-empty. If the \code{type} is \code{"folder"}, the function removes any trailing slashes from \code{name}
and checks if the directory exists. If the directory does not exist, it is created using \code{dir.create} with the specified \code{recursive}
setting. A message is printed to indicate the creation of the folder. Finally, it asserts that the directory exists and is executable.
}
\examples{
\dontrun{
# Create a folder "data" if it does not already exist
createIfNotExist("folder", "data")
}

}
