% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Private.R
\name{formatDecimalWithComma}
\alias{formatDecimalWithComma}
\title{Format a Number with Commas and Specified Decimal Places}
\usage{
formatDecimalWithComma(number, decimalPlaces = 1, round = TRUE)
}
\arguments{
\item{number}{A numeric value to be formatted.}

\item{decimalPlaces}{An integer specifying the number of digits to display after the decimal point.
Defaults to \code{1}.}

\item{round}{A logical value indicating whether the decimal portion should be rounded. If \code{FALSE},
the decimal part will be truncated instead. Defaults to \code{TRUE}.}
}
\value{
A character string representing the formatted number with commas as thousand separators and
a period separating the integer and decimal parts.
}
\description{
This function formats a numeric value into a string by inserting commas as thousand separators
and formatting the decimal portion to a specified number of digits. It can either round or truncate
the decimal part based on the \code{round} parameter.
}
\details{
The function separates the number into its integer and decimal components. The integer part is formatted
with commas using \code{formatC}, while the decimal part is either rounded or truncated according to the
\code{decimalPlaces} parameter. The decimal portion is then extracted from the formatted string and concatenated
with the formatted integer part to produce the final output.
}
\examples{
# Example with rounding (default)
formatDecimalWithComma(1234567.8912)
# Might return "1,234,567.9"

# Example with truncation and two decimal places
formatDecimalWithComma(1234567.8912, decimalPlaces = 2, round = FALSE)
# Might return "1,234,567.89"

}
