% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCohortPathways.R
\name{plotCohortPathways}
\alias{plotCohortPathways}
\title{Plot cohort pathway.}
\usage{
plotCohortPathways(
  resultsFolder,
  exportFolder,
  pathwayAnalysisGenerationId = NULL,
  targetCohortId = NULL,
  height = 400,
  width = "100\%"
)
}
\arguments{
\item{resultsFolder}{The folder where the output from executeCohortPathways is located. If provided, then
the function will look for pathwayAnalysisCodes.csv and pathwayAnlaysisPaths.csv. Ignored
if pathwayAnalysisCodes or pathwayAnalysisPaths are provided.}

\item{exportFolder}{The folder where the output will be exported to. If this folder
does not exist it will be created.}

\item{pathwayAnalysisGenerationId}{(Optional) Generation id of pathway. This uniquely identifies every generation execution.}

\item{targetCohortId}{(Optional) The cohort id of the target cohort.}

\item{height}{(Optional) height and width of sunburst htmlwidget containing div specified in any valid CSS size unit.}

\item{width}{(Optional) height and width of sunburst htmlwidget containing div specified in any valid CSS size unit.}
}
\description{
Uses the output of executeCohortPathways and using sunburstD3 produces sunburst plots
}
\examples{
\dontrun{

executeCohortPathways(
  resultsFolder = "results",
  pathwayAnalysisGenerationId = 1234,
  targetCohortId = 1
)
}

}
