% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortPathways.R
\docType{package}
\name{CohortPathways-package}
\alias{CohortPathways}
\alias{CohortPathways-package}
\title{CohortPathways: Create Pathways from Target to Event Cohorts}
\description{
Software tool designed to compute the temporal relationship defined as pathways between any two instantiated cohorts. The cohorts are input as Target and event cohorts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OHDSI/CohortPathways}
  \item Report bugs at \url{https://github.com/OHDSI/CohortPathways/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Authors:
\itemize{
  \item Chris Knoll \email{cknoll@ohdsi.org}
}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
