% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CohortPathways.R
\docType{package}
\name{CohortPathways-package}
\alias{CohortPathways}
\alias{CohortPathways-package}
\title{CohortPathways: Cohort Pathways to create pathways for Target and Event cohorts}
\description{
An R package that creates new cohort(s) from previously instantiated cohorts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/CohortPathways/}
  \item \url{https://github.com/OHDSI/CohortPathways}
  \item Report bugs at \url{https://github.com/OHDSI/CohortPathways/issues}
}

}
\author{
\strong{Maintainer}: Gowtham Rao \email{rao@ohdsi.org}

Other contributors:
\itemize{
  \item Observational Health Data Science and Informatics [copyright holder]
}

}
\keyword{internal}
