% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSunburst.R
\name{createPathwaySankey}
\alias{createPathwaySankey}
\title{Create a Sankey Plot for Treatment Pathways}
\usage{
createPathwaySankey(cpResults, generationSet, maxPaths = 5, minCount = 10)
}
\arguments{
\item{cpResults}{A list containing the results from CohortPathway analysis.}

\item{generationSet}{A data frame containing information about event cohorts.}

\item{maxPaths}{The maximum number of steps (depth) to include in the plot.}

\item{minCount}{The minimum count of a pathway to be included in the visualization.}
}
\value{
A Sankey network plot created with the networkD3 package.
}
\description{
This function creates an interactive Sankey diagram to visualize the flow of
patients through different treatment pathways.
}
\examples{
\dontrun{
library(CohortPathways)
# Assuming cpResults and generationSet are loaded
sankeyPlot <- createPathwaySankey(
  cpResults = cpResults,
  generationSet = generationSet,
  maxPaths = 5,
  minCount = 10
)
sankeyPlot
}
}
