% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPlots.R
\name{createPathwaySunburst}
\alias{createPathwaySunburst}
\title{Create a Sunburst Plot for Treatment Pathways}
\usage{
createPathwaySunburst(
  cpResults,
  generationSet,
  nPaths = 3,
  minCount = 5,
  plotWidth = "80\%",
  plotHeight = 600
)
}
\arguments{
\item{cpResults}{A list containing the results from CohortPathway analysis.
Must include 'pathwaysAnalysisPathsData' and 'isCombo' data frames.}

\item{generationSet}{A data frame containing information about cohorts
that will be used to generate descriptive names for the events in the diagram and target names of the plot}

\item{nPaths}{Integer specifying the maximum number of steps to include in the plot.}

\item{minCount}{Integer specifying the minimum count value for a path to be included.}
}
\value{
An HTML widget object containing the interactive sunburst plot.
}
\description{
This function creates an interactive sunburst plot to visualize treatment pathways
from CohortPathway analysis results. The plot shows the hierarchical structure of
treatment sequences, with each ring representing a step in the pathway.
}
\examples{

\dontrun{
library(CohortPathway)
sunburstPlot <- CohortPathways::createPathwaySunburst(cpResults, generationSet)
}
}
