% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSunburst.R
\name{createPathwaySunburst}
\alias{createPathwaySunburst}
\title{Create a Sunburst Plot for Treatment Pathways}
\usage{
createPathwaySunburst(
  cpResults,
  generationSet,
  nPaths = 3,
  minCount = 5,
  debug = TRUE
)
}
\arguments{
\item{cpResults}{A list containing the results from CohortPathway analysis.
Must include 'pathwaysAnalysisPathsData' and 'isCombo' data frames.}

\item{generationSet}{A data frame containing information about event cohorts
that will be used to generate descriptive names for the events in the diagram.}

\item{nPaths}{Integer specifying the maximum number of steps to include in the plot.}

\item{minCount}{Integer specifying the minimum count value for a path to be included.}

\item{debug}{Logical indicating whether to print debug information.}
}
\value{
An HTML widget object containing the interactive sunburst plot.
}
\description{
This function creates an interactive sunburst plot to visualize treatment pathways
from CohortPathway analysis results. The plot shows the hierarchical structure of
treatment sequences, with each ring representing a step in the pathway.
}
